refactor(api): DRY search logic into helpers (buildSearchCondition & filterAdvocatesMock)
